backend/requirements.txt
# Flask Backend Dependencies — Mie Gacoan ABSA Fullstack
# ========================================================

# Core Framework
flask>=3.0.0
flask-cors>=4.0.0
flask-restful>=0.3.10

# Autentikasi JWT
flask-jwt-extended>=4.6.0
PyJWT>=2.8.0

# Database PostgreSQL + ORM
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.25

# Password Hashing
bcrypt>=4.1.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# NLP & ML (untuk Live Prediction)
scikit-learn>=1.4.0
nltk>=3.8.1
Sastrawi>=1.0.1
joblib>=1.3.0

# Environment & Utilities
python-dotenv>=1.0.0
marshmallow>=3.20.0

# Production WSGI Server
gunicorn>=21.2.0
